requirements.txt
mysql-connector-python==8.0.29
protobuf==4.23.2
